+++ NDTensors/src/NamedDimsArrays/ext/NamedDimsArraysTensorAlgebraExt/src/NamedDimsArraysTensorAlgebraExt.jl
@@ -1,3 +1,6 @@
 module NamedDimsArraysTensorAlgebraExt
 include("contract.jl")
+include("qr.jl")
+include("eigen.jl")
+include("svd.jl")
 end
