+++ NDTensors/src/NamedDimsArrays/ext/NamedDimsArraysTensorAlgebraExt/test/runtests.jl
@@ -1,8 +1,9 @@
 @eval module $(gensym())
-using Test: @test, @testset
+using Test: @test, @testset, @test_broken
 using NDTensors.NamedDimsArrays: named, unname
 using NDTensors.TensorAlgebra: TensorAlgebra
-@testset "NamedDimsArraysTensorAlgebraExt (eltype=$(elt))" for elt in (
+using LinearAlgebra: qr
+@testset "NamedDimsArraysTensorAlgebraExt contract (eltype=$(elt))" for elt in (
   Float32, ComplexF32, Float64, ComplexF64
 )
   i = named(2, "i")
@@ -13,5 +14,16 @@ using NDTensors.TensorAlgebra: TensorAlgebra
   na_dest = TensorAlgebra.contract(na1, na2)
   @test eltype(na_dest) === elt
   @test unname(na_dest, (i, k)) ≈ unname(na1) * unname(na2)
+end
+@testset "NamedDimsArraysTensorAlgebraExt QR (eltype=$(elt))" for elt in (
+  Float32, ComplexF32, Float64, ComplexF64
+)
+  di = 2
+  dj = 2
+  i = named(di, "i")
+  j = named(dj, "j")
+  na = randn(elt, i, j)
+  @test_broken error("QR not implemented yet")
+  # q, r = qr(na)
 end
 end
