+++ NDTensors/src/NamedDimsArrays/src/NamedDimsArrays.jl
@@ -1,4 +1,6 @@
 module NamedDimsArrays
+include("BaseExtension/BaseExtension.jl")
+
 include("traits.jl")
 include("abstractnamedint.jl")
 include("abstractnamedunitrange.jl")
@@ -7,6 +9,7 @@ include("namedint.jl")
 include("namedunitrange.jl")
 include("nameddimsarray.jl")
 include("constructors.jl")
+include("tensoralgebra.jl")
 
 # Extensions
 include("../ext/NamedDimsArraysAdaptExt/src/NamedDimsArraysAdaptExt.jl")
